requirements.txt
streamlit>=1.33.0
openai>=1.30.0
langchain>=0.2.0
langchain-openai>=0.1.6
python-dotenv>=1.0.1